backend/requirements.txt
absl-py==2.4.0
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
distro==1.9.0
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.61.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.32
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
onnxruntime==1.23.2
opencv-contrib-python==4.13.0.90
opencv-python==4.13.0.90
packaging==26.0
pillow==12.1.0
polars==1.37.1
polars-runtime-32==1.37.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scipy==1.17.0
setuptools==80.10.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
sympy==1.14.0
tenacity==9.1.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.9
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
supabase==2.10.0
flask==3.1.0
flask-cors==5.0.0
